complete cud and fix bug

+++ controllers/article.controllers.js
@@ -1,6 +1,6 @@
 const slugify = require("slugify");
 const ArticleModel = require("../models/article");
-const { isAdmin } = require("../uttils/functions");
+const { isAdmin, unlinkCover } = require("../uttils/functions");
 
 class ArticleController {
   async createArticle(req, res, next) {
@@ -29,15 +29,59 @@ class ArticleController {
       next(error);
     }
   }
-  async deleteArticle(req, res, next) {
+  async removeArticle(req, res, next) {
     try {
       const slug = req?.params?.slug;
-      const id = req.user._id;
-      const role = req.user.role === "admin" ? true : false;
+      const userId = req.user._id;
 
-      const result = ArticleModel.findOneAndDelete({
-        slug,
-        author: role ? id : null,
+      const article = await ArticleModel.findOne(
+        isAdmin(req) ? { slug } : { slug, author: userId }
+      );
+      if (!article) throw { status: 404, message: "Article not founded" };
+      unlinkCover(article.cover);
+
+      const result = await ArticleModel.findOneAndDelete(
+        isAdmin(req) ? { slug } : { slug, author: userId }
+      );
+      if (result.deletedCount == 0)
+        throw { status: 400, message: "There is a problem to remove article" };
+      return res.status(200).json({
+        status: "success",
+        message: "article successfully deleted!",
+      });
+    } catch (error) {
+      next(error);
+    }
+  }
+  async editArticle(req, res, next) {
+    try {
+      const slug = req?.params?.slug;
+      const userId = req.user._id;
+      const { title, desc, status } = req.body;
+      const newSlug = slugify(title);
+      const filename = req.file?.filename;
+
+      const article = await ArticleModel.findOne(
+        isAdmin(req) ? { slug } : { slug, author: userId }
+      );
+      if (!article) throw { status: 404, message: "Article not founded" };
+
+      unlinkCover(article.cover);
+
+      const result = await ArticleModel.findOneAndUpdate(
+        isAdmin(req) ? { slug } : { slug, author: userId },
+        {
+          $set: { title, desc, status, slug: newSlug, cover: filename },
+        }
+      );
+      if (result.modifiedCount == 0)
+        throw {
+          status: 400,
+          message: "There was a problem updating the article",
+        };
+      return res.status(200).json({
+        status: "success",
+        message: "Updated",
       });
     } catch (error) {
       next(error);
@@ -46,11 +90,9 @@ class ArticleController {
   async getArticle(req, res, next) {
     try {
       const slug = req?.params?.slug;
-      const role = req.user.role === "admin" ? true : false;
-      const article = await ArticleModel.findOne({
-        slug,
-        author: role ? req.user._id : null,
-      });
+      const article = await ArticleModel.findOne(
+        (await isAdmin(req)) ? { slug } : { status: "published", slug }
+      );
       if (!article)
         throw { status: 404, message: "Coudln't find your article" };
       // increment views
